添加获取messages在get之后不删除的参数

+++ src/controller/messages.js
@@ -67,7 +67,7 @@ router.get('/getAllUnreadMessagesWithOneUser', function*() {
   this.body = messages;
 
   //消息被提取之后删除
-  if (process.env.NODE_ENV === 'development') {
+  if (process.env.NODE_ENV === 'development' && this.query.isSave !== 'true') {
     yield models.Message.destroy({
       where: {
         UserId: this.query.userId,
@@ -94,7 +94,7 @@ router.get('/getAllUnreadMessages', function*() {
   this.body = messages.map((message) => message.toJSON());
 
   //消息被提取之后删除
-  if (process.env.NODE_ENV === 'development') {
+  if (process.env.NODE_ENV === 'development' && this.query.isSave !== 'true') {
     yield models.Message.destroy({
       where: {
         TargetUserId: this.session.user.id,
@@ -140,7 +140,7 @@ router.get('/getAllMessages', function*() {
   this.body = messages.map((message) => message.toJSON());
 
   //消息被提取之后删除
-  if (process.env.NODE_ENV === 'development') {
+  if (process.env.NODE_ENV === 'development' && this.query.isSave !== 'true') {
     yield models.Message.destroy({
       where: {
         id: {
